+++ pages/_app.tsx
@@ -3,8 +3,9 @@ import '../styles/globals.css';
 import Head from 'next/head';
 
 import MDXProvider from "../components/MDXProvider";
+import { AppProps } from 'next/app';
 
-function MyApp({ Component, pageProps }) {
+function MyApp({ Component, pageProps }: AppProps) {
   return <>
     <Head>
       <link rel="icon" href="/favicon.png" />
