+++ pages/api/hello.tsx
@@ -1,0 +1,6 @@
+// Next.js API route support: https://nextjs.org/docs/api-routes/introduction
+import { NextApiRequest, NextApiResponse } from 'next'
+
+export default function handler(req: NextApiRequest, res: NextApiResponse<{name: string}>) {
+  res.status(200).json({ name: 'John Doe' })
+}
